+++ packages/blade/src/components/Input/BaseInput/StyledBaseInput.native.tsx
@@ -269,7 +269,7 @@ const _StyledBaseInput: React.ForwardRefRenderFunction<
       {...accessibilityProps}
     >
       <Text
-        type={props.value ? 'subtle' : 'placeholder'}
+        type={props.value && !isDisabled ? 'subtle' : 'placeholder'}
         truncateAfterLines={1}
         textAlign={props.textAlign}
       >